Restructure Program again
